Improve integration tests for sync files
